Improvements for data management

+++ backend/hmap/app/goes/goes.py
@@ -21,8 +21,11 @@ def get_s3_folder_prefix(date, hour):
 
 def get_s3_link(date_str, hour, bucket):
   date = datetime.datetime.strptime(date_str, "%m/%d/%Y")
-  for key in s3.list_objects(Bucket=bucket, Prefix=get_s3_folder_prefix(date, hour))['Contents']:
-      return key['Key']
+  for key in s3.list_objects(Bucket=bucket, Prefix=get_s3_folder_prefix(date, hour)):
+    if 'Contents' in key:
+      return key['Contents']['Key']
+    else:
+      return None
 
 def get_data(lat, lon, mask, channels):
   res = []
@@ -44,21 +47,30 @@ def lat_lon_to_geojson(coordinates):
 def getFireFromGoesByBucket(bucket, date, hour):
     channels = [10,11,12,13,14,15,30,31,32,33,34,35]
     filename = "tmp/" + bucket + date.replace("/","-") + hour
-    if not os.path.isfile(filename + ".geojson"):
-      s3.download_file(bucket, get_s3_link(date, hour, bucket), filename + ".nc")
-      ds = Dataset(filename + ".nc", 'r')
-      (lat, lon) = goes_to_latlon(ds)
-      mask = ds.variables['Mask'][:]
-      indian = get_data(lat, lon, mask, channels)
-      ds.close()
-      ds = None
-      data = lat_lon_to_geojson(indian)
-      os.remove(filename + ".nc")
-      with open(filename + ".geojson", 'w') as outfile:
-        json.dump(data, outfile)
-    else:
-      f = open(filename + ".geojson", "r")
-      data = json.loads(f.read())
+    s3_link = None
+    while s3_link == None and int(hour) > 0:
+      s3_link = get_s3_link(date, hour, bucket)
+      hour = int(hour) - 1
+      if not os.path.isfile(filename + ".geojson"):
+        if s3_link is not None:
+          s3.download_file(bucket, s3_link, filename + ".nc")
+          ds = Dataset(filename + ".nc", 'r')
+          (lat, lon) = goes_to_latlon(ds)
+          mask = ds.variables['Mask'][:]
+          indian = get_data(lat, lon, mask, channels)
+          ds.close()
+          ds = None
+          data = lat_lon_to_geojson(indian)
+          os.remove(filename + ".nc")
+          with open(filename + ".geojson", 'w') as outfile:
+            json.dump(data, outfile)
+        else:
+          data = {
+            'features': []
+          }
+      else:
+        f = open(filename + ".geojson", "r")
+        data = json.loads(f.read())
     return data
 
 def getFireFromGoes():
